ExamApp/app/requirements.txt
psycopg2-binary==2.9.1
gunicorn==20.1.0
Django==3.2.6
Pillow==2.1.0
django-cors-headers
djangorestframework
python-dotenv